// backend/src/modules/agents/types/lll-models.enum.ts
export enum LLMModels {
  GEMINI_2_5_FLASH = 'models/gemini-2.5-flash',
  GEMINI_2_5_PRO = 'models/gemini-2.5-pro',
  GEMINI_3_FLASH = 'models/gemini-3-flash-preview',
  GEMINI_3_PRO = 'models/gemini-3-pro-preview',
  GEMINI_FLASH_LATEST = 'models/gemini-flash-latest',
  GEMINI_PRO_LATEST = 'models/gemini-pro-latest',
}